Ignore inactive services
Added logic to look up the service in the cattle api and ignore it if it is stopped/inactive.
This required calling into the cattle api for looking up the service state.
It didn't make sense for the metadata wrapper to depend on the cattle wrapper so I broke logic out into a new scheduler.